+++ phis2-ws/src/main/java/opensilex/service/resource/brapi/GermplasmResourceService.java
@@ -15,115 +15,35 @@ import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import java.util.ArrayList;
-import java.util.List;
-import javax.servlet.http.HttpServletRequest;
-import javax.validation.Valid;
 import javax.validation.constraints.Min;
-import javax.ws.rs.Consumes;
 import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
-import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import opensilex.service.configuration.DefaultBrapiPaginationValues;
 import opensilex.service.configuration.GlobalWebserviceValues;
 import opensilex.service.dao.AccessionDAO;
 import opensilex.service.documentation.DocumentationAnnotation;
-import opensilex.service.documentation.StatusCodeMsg;
 import opensilex.service.model.Accession;
 import opensilex.service.resource.ResourceService;
-import opensilex.service.resource.dto.germplasm.GermplasmDTO;
-import opensilex.service.resource.dto.germplasm.AccessionPostDTO;
+import opensilex.service.resource.dto.accession.GermplasmDTO;
 import opensilex.service.resource.validation.interfaces.URL;
 import opensilex.service.result.ResultForm;
-import opensilex.service.utils.POSTResultsReturn;
 import opensilex.service.view.brapi.Status;
-import opensilex.service.view.brapi.form.AbstractResultForm;
-import opensilex.service.view.brapi.form.ResponseFormPOST;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Accession resource service
+ * Germplasm resource service
  * @author Alice Boizet <alice.boizet@inra.fr>
  */
 @Api("/brapi/v1/germplasm")
 @Path("/brapi/v1/germplasm")
 public class GermplasmResourceService extends ResourceService {
     final static Logger LOGGER = LoggerFactory.getLogger(GermplasmResourceService.class);
-    
-    /**
-     * Inserts germplasm in the storage.
-     * @param germplasm list of germplasm to insert.
-     * @example
-     * {
-     *      
-     * }
-     * @param context
-     * @return the post result with the errors or the uri of the inserted germplasm
-     */
-    @POST
-    @ApiOperation(value = "Post Accession,",
-                  notes = "Register new Accession in the database")
-    @ApiResponses(value = {
-        @ApiResponse(code = 201, message = "Accessions saved", response = ResponseFormPOST.class),
-        @ApiResponse(code = 400, message = DocumentationAnnotation.BAD_USER_INFORMATION),
-        @ApiResponse(code = 401, message = DocumentationAnnotation.USER_NOT_AUTHORIZED),
-        @ApiResponse(code = 500, message = DocumentationAnnotation.ERROR_SEND_DATA)
-    })
-    @ApiImplicitParams({
-        @ApiImplicitParam(name = GlobalWebserviceValues.AUTHORIZATION, required = true,
-                dataType = GlobalWebserviceValues.DATA_TYPE_STRING, paramType = GlobalWebserviceValues.HEADER,
-                value = DocumentationAnnotation.ACCES_TOKEN,
-                example = GlobalWebserviceValues.AUTHENTICATION_SCHEME + " ")
-    })
-    @Consumes(MediaType.APPLICATION_JSON)
-    @Produces(MediaType.APPLICATION_JSON)
-    public Response post(
-        @ApiParam (value = DocumentationAnnotation.GERMPLASM_POST_DEFINITION) @Valid ArrayList<AccessionPostDTO> accession,
-        @Context HttpServletRequest context) throws Exception {
-        AbstractResultForm postResponse = null;
-        
-        if (accession != null && !accession.isEmpty()) {
-            AccessionDAO accessionDAO = new AccessionDAO();
-            
-            accessionDAO.user = userSession.getUser();
-            
-            POSTResultsReturn result = accessionDAO.checkAndInsert(accessionPostDTOsToAccession(accession));
-            
-            if (result.getHttpStatus().equals(Response.Status.CREATED)) {
-                postResponse = new ResponseFormPOST(result.statusList);
-                postResponse.getMetadata().setDatafiles(result.getCreatedResources());
-            } else if (result.getHttpStatus().equals(Response.Status.BAD_REQUEST)
-                    || result.getHttpStatus().equals(Response.Status.OK)
-                    || result.getHttpStatus().equals(Response.Status.INTERNAL_SERVER_ERROR)) {
-                postResponse = new ResponseFormPOST(result.statusList);
-            }
-            return Response.status(result.getHttpStatus()).entity(postResponse).build();
-        } else {
-            postResponse = new ResponseFormPOST(new Status(StatusCodeMsg.REQUEST_ERROR, StatusCodeMsg.ERR, "Empty accession(s) to add"));
-            return Response.status(Response.Status.BAD_REQUEST).entity(postResponse).build();
-        }
-    }
-    
-    /**
-     * Generates a germplasm list from a given list of AccessionPostDTO
-     * @param accessionDTOs
-     * @return the list of sensors
-     */
-    private List<Accession> accessionPostDTOsToAccession(List<AccessionPostDTO> accessionDTOs) throws Exception {
-        ArrayList<Accession> accessions = new ArrayList<>();
-        
-        for (AccessionPostDTO accessionDTO : accessionDTOs) {
-            accessions.add(accessionDTO.createObjectFromDTO());
-        }
-        
-        return accessions;
-    }
     
     @GET
     @ApiOperation(value = "Get all germplasm corresponding to the search params given",
